Task #11 - Refactor JWT class location - bumped the JSR-375/Soteria version to 1.0 - updated the Liquibase profile to use localhost instead of the container IP address - moved the classes from the jwt package into the control package
