+++ servicetwo/pom.xml
@@ -44,15 +44,10 @@
         </dependency>
 
         <!-- JSR 375 Security -->
-<!--        <dependency>
-            <groupId>javax.security.enterprise</groupId>
-            <artifactId>javax.security.enterprise-api</artifactId>
-            <version>1.0-b11</version>
-        </dependency>-->
         <dependency>
             <groupId>org.glassfish.soteria</groupId>
             <artifactId>javax.security.enterprise</artifactId>
-            <version>1.0-b11</version>
+            <version>1.0</version>
         </dependency>
     </dependencies>
